renamed classes and refactored obvious methods
