+++ src/menu/sub_menus/CustomerMenu.java
@@ -35,14 +35,6 @@ public class CustomerMenu extends Menu implements DBStandardQueries {
         requestHandler.printQueryResult(query, printColumnNames, db_dependencies.CUSTOMER_COLUMNS);
     }
 
-//    @Override
-//    public void insertToTable(DB_QueryEditingHandler editingHandler,DB_QueryRequestHandler requestHandler, UI ui) {
-//        String query = "INSERT INTO customer_info " + "(" +
-//                getSpecifiedTableNames(1, 2, 3, 4, 5, 6, 7, 8) +
-//                ")" + " VALUES (" + getValuesToInsert(ui) + ");";
-//
-//        editingHandler.insertQuery(query);
-//    }
 
     @Override
     public void insertToTable(DB_QueryEditingHandler editingHandler,DB_QueryRequestHandler requestHandler, UI ui) {
